Feature/serial (#62)
* Serial port work for timeouts

* Serial port work - requires Nuttx 3.9 update

* Serial port fixes

* Add ability to set clock

* PWM channel bug fix

* Update azure-pipelines.yml

* Add an Overrun event to serial port

+++ source/Meadow.Core/Interop/Interop.timespec.cs
@@ -10,8 +10,8 @@ namespace Meadow.Core
 #pragma warning disable 0649
             public struct timespec
             {
-                public int tv_sec;
-                public int tv_nsec;
+                public long tv_sec;
+                public long tv_nsec;
             }
 #pragma warning disable 0649
